% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alroy_3timers.R
\name{Alroy_3timers}
\alias{Alroy_3timers}
\title{Function to calculate Alroy's three-timer rates of macroevolutionary change}
\usage{
Alroy_3timers(Obs, dts = rep(1, dim(Obs)[2]))
}
\arguments{
\item{Obs}{matrix of observations (taxa by time)}

\item{dts}{vector of interval duration (defaults to 1)}
}
\value{

}
\description{
Function to calculate Alroy's three-timer rates of macroevolutionary change
}
