% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkchains.R
\name{checkchains}
\alias{checkchains}
\title{A quick function to plot chains and calc ESS's for drivers and mean effects.}
\usage{
checkchains(cmrfit)
}
\arguments{
\item{cmrfit}{A fitted model using MCMC_CMR}
}
\description{
A quick function to plot chains and calc ESS's for drivers and mean effects.
}
