% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colbottom.R
\name{colbottom}
\alias{colbottom}
\title{Simple function to add colorbars at the lower end of a plot. Useful for coloring
different stages when plotting rates over time.}
\usage{
colbottom(stages)
}
\arguments{
\item{stages}{- data.frame of stages used. The function uses the min_ma, max_ma and the strings in color. See GSA_timescale.}
}
\description{
Simple function to add colorbars at the lower end of a plot. Useful for coloring
different stages when plotting rates over time.
}
