% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contMCMC_CMR.R
\name{contMCMC_CMR}
\alias{contMCMC_CMR}
\title{Function to continue sampling a fit, without adaptation.}
\usage{
contMCMC_CMR(fit, x0 = NULL, niter = 1e+05, nthin = 10, ...)
}
\arguments{
\item{fit}{is the output from \link{MCMC_CMR}}

\item{x0}{is the initial parameter values for continued sampling. If not given taken as the last parameter set from \emph{fit}}

\item{niter}{is the number of iterations}

\item{nthin}{is the thinning of the chain}
}
\value{
a fit structure of class \emph{CMR_fit}
}
\description{
After an initial trial with \link{MCMC_CMR} which has seemingly reasonable mixing, this function
can be used to continue sampling using the estimated covariance proposal matrix.
}
