% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doBML.R
\name{doBML}
\alias{doBML}
\title{Function to approximate Bayesian model probability using importance sampling.
The output logBML is approximating the Bayesian factor integral, and the support for one model over another, given the same data (i.e. the Bayes factor), can be approximated by exp(logBML[1] - logBML[2])}
\usage{
doBML(myf, chain, ndraws = 10000)
}
\arguments{
\item{myf}{a Bayesian proportional probability density function}

\item{chain}{a chain from an MCMC fit. Exclude burning before submitting.}

\item{ndraws}{number of samples to use to approximate the bayesian model probability}
}
\description{
Function to approximate Bayesian model probability using importance sampling.
The output logBML is approximating the Bayesian factor integral, and the support for one model over another, given the same data (i.e. the Bayes factor), can be approximated by exp(logBML[1] - logBML[2])
}
