% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESS.R
\name{ESS}
\alias{ESS}
\title{Function to extract effective sample sizes from a MCMC-CMR fit.
Calculated by using effectiveSize from the package coda.}
\usage{
ESS(cmrfit)
}
\arguments{
\item{fit}{A fit from MCMC_CMR. Extracts samples from fit$Chain}
}
\value{
ESSs - an array of Effective sample size for each variable.
}
\description{
Function to extract effective sample sizes from a MCMC-CMR fit.
Calculated by using effectiveSize from the package coda.
}
