% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Foote_percap.R
\name{Foote_percap}
\alias{Foote_percap}
\title{Function to calculate Foote's rates of macroevolutionary change.}
\usage{
Foote_percap(Obs, dts)
}
\arguments{
\item{Obs}{matrix of observations (species by time)}

\item{dts}{vector of interval durations}
}
\value{
p_hat,q_hat, N_b_t, N_t,N_b
}
\description{
Function to calculate Foote's rates of macroevolutionary change.
}
