% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInit.R
\name{getInit}
\alias{getInit}
\title{A function to generate sensible initial values for MCMC estimation.}
\usage{
getInit(cmrModel)
}
\arguments{
\item{cmrModel}{A CMR model structure.}
}
\value{
The function returns relatively decent initial values for MCMC estimation.
}
\description{
getInits generates initial values for a complex CMR model. Essentially the global parameters (mean rates) are estimated optimizing the probability density function, assuming all random effects are 0. Then each random effect is estimated independently, assuming all other random effects are 0. This is not proper optimization, but work reasonably well for initial values.
}
\examples{
mod1 <- make.BayesCMR(Obs)
x0 <- getInit(mod1)
fit <- MCMC_CMR(mod1,x0=x0)
}
