% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRates.R
\name{getRates}
\alias{getRates}
\title{Extract macroevolutionary rates from CMR model fit.}
\usage{
getRates(fit, ns = 250)
}
\arguments{
\item{fit}{A fit returned from \link{MCMC_CMR}}

\item{ns}{number of samples used in chain.}
}
\description{
Extract macroevolutionary rates from CMR model fit.
}
