% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPDI.R
\name{HPDI}
\alias{HPDI}
\title{Function to extract highest posterior density intervals.
HPDI corresponds to the 'shortest' intervals that contain p fraction of the samples.}
\usage{
HPDI(x, p = 0.95)
}
\arguments{
\item{x}{Samples of one parameter or a CMR_fit.}

\item{p}{probability. Defaults to 0.95.}
}
\value{
returns hpdi and exact probability of the interval, either for the given sample or for each parameter in a CMR_fit.
}
\description{
Function to extract highest posterior density intervals.
HPDI corresponds to the 'shortest' intervals that contain p fraction of the samples.
}
