% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_BayesCMR_2clades.R
\name{make_BayesCMR_2clades}
\alias{make_BayesCMR_2clades}
\title{Function to construct two clade CMR-model.}
\usage{
make_BayesCMR_2clades(Obs1, Obs2, dts = rep(1, dim(Obs1)[2]),
  intSpec = T, intExt = T, SpecTS1 = NULL, ExtTS1 = NULL,
  SmpTS1 = NULL, SpecTS2 = NULL, ExtTS2 = NULL, SmpTS2 = NULL, ...)
}
\arguments{
\item{Obs1}{observations clade 1: a matrix with size \emph{number of taxa} by \emph{number of intervals (n)}. Each taxa has a row with 0's (unobserved) and 1's (observed) for each interval in the analysis. Oldest interval is first column.}

\item{Obs2}{observations clade 1: a matrix with size \emph{number of taxa} by \emph{number of intervals (n)}. Each taxa has a row with 0's (unobserved) and 1's (observed) for each interval in the analysis. Oldest interval is first column.}

\item{dts}{a vector of interval durations. Defaults to series of 1's if not given.}

\item{intSpec}{TRUE/FALSE on whether or not diversities of both clades affect speciation rates of both.}

\item{SpecTS1/2}{drivers for speciation rates for clades 1/2}

\item{ExtTS1/2}{drivers for extinction rates for clades 1/2}

\item{SmpTS1/2}{drivers for sampling rates}
}
\value{
this function yields whatnow?
}
\description{
This function generates a structure containing the necessary functions for a
CMR analysis of two clades. In addition to drivers affecting both clades, their rates of
speciation and extinction can depend on the estimated diversity of both clades.
}
\details{
Both clades must be observed at each interval and these must be the same for the two clades.
}
\examples{
M1 <- make_BayesCMR_2clades(Obs1,Obs2,dts=rep(2,5))



}
