% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_unvd.R
\name{make_unvd}
\alias{make_unvd}
\title{Generate datavectors used in Pradel likelihood calculations.}
\usage{
make_unvd(Obs)
}
\arguments{
\item{Obs}{a matrix with size \emph{number of taxa} by \emph{number of intervals}. Each taxa has a row with 0's (unobserved) and 1's (observed) for each interval in the analysis.}
}
\value{
a list containing \emph{Obs}, \emph{u,n,v} and \emph{d}.
}
\description{
This function takes an observation matrix \emph{Obs} and generates the vectors u,n,v needed to calculate the Pradel likelihood. Used internally by other functions.
}
