% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cmr_simulation.R
\name{plot.cmr_simulation}
\alias{plot.cmr_simulation}
\title{Plots simulation output in Compadre.}
\usage{
\method{plot}{cmr_simulation}(Sim1, drawplot = T)
}
\arguments{
\item{Sim1}{Output from sim_bd_func}
}
\description{
Produces simple figure with the true number of taxa and the unique number of species observed within each interval.
}
