% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDrivers.R
\name{plotDrivers}
\alias{plotDrivers}
\title{Function to make simple plot of estimated drivers (and intercepts) in CMR model fit.}
\usage{
plotDrivers(cmrfit, nsamp = 250, spec = T, ext = T, samp = T)
}
\arguments{
\item{cmrfit}{A model fit, as output from \link{MCMC_CMR}}
}
\description{
Function to make simple plot of estimated drivers (and intercepts) in CMR model fit.
}
