% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfun_internal.R
\name{plotfun_internal}
\alias{plotfun_internal}
\title{internal plottingfunction
export}
\usage{
plotfun_internal(spec_smps, ext_smps, smp_smps, bnds, tmpl,
  stages = NULL, quantiles = c(0.025, 0.5, 0.975), main = "")
}
\description{
internal plottingfunction
export
}
