% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotN_sim.R
\name{plotN_sim}
\alias{plotN_sim}
\title{Plots simulation output in Compadre.}
\usage{
plotN_sim(Sim1)
}
\arguments{
\item{Sim1}{Output from sim_bd_func}
}
\description{
Produces simple figure with the true number of taxa and the unique number of species observed within each interval.
}
