% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRates.R
\name{plotRates}
\alias{plotRates}
\title{Plot macroevolutionary rates from a CMR model fit.}
\usage{
plotRates(fit_1, max_ma = sum(fit_1$Model$dts), botcols = NULL,
  logax = T, qs = c(0.025, 0.5, 0.975))
}
\arguments{
\item{fit_1}{A model fit, as output from \link{MCMC_CMR}}

\item{max_ma}{(optional) a maximum age for the oldest interval.}

\item{botcols}{(optional) a list of colors for each interval. If supplied the lower part of the lower panel is colored.}
}
\description{
Plot macroevolutionary rates from a CMR model fit.
}
