% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pradel_unvd_gam.R
\name{pradel_unvd_gam}
\alias{pradel_unvd_gam}
\title{Internal function used to calculate the likelihood of the Pradel CMR model using the seniority as input.}
\usage{
pradel_unvd_gam(gam, ext, p, u, n, v, d, minval = 1e-05)
}
\arguments{
\item{gam}{Seniority in the Pradel model. Either one value or one for each of the n-1 last intervals.}

\item{ext}{Extinction probabilities. Same size as fec.}

\item{p}{Sampling probabilities.}

\item{u, n, v, d}{Values for number observations, each of length equal to number of intervals. u - number of taxa first observed, n - number of taxa observed, v - number of taxa observed for last time. d should be vector of 0's.}

\item{minval}{Minimum probabilities used for calculations. No probabilities are allowed be be smaller than minval or larger than 1-minval. Defaults to 1e-5.}
}
\value{
Returns $LogL (log likelihood), and $Ll.numerator and $Ll.denominator.
}
\description{
Internal function used to calculate the likelihood of the Pradel CMR model using the seniority as input.
}
