% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob2rate.R
\name{prob2rate}
\alias{prob2rate}
\title{Function transform a probability into a rate for an interval}
\usage{
prob2rate(p, dt)
}
\arguments{
\item{p}{probability of an event}

\item{dt}{duration of interval}
}
\value{
The function returns the transformed rate -log(1-p)/dt
}
\description{
Function transform a probability into a rate for an interval
}
