% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_ext.R
\name{prob_ext}
\alias{prob_ext}
\title{A function to calculate the probability of extinction for a paraclade.}
\usage{
prob_ext(m, l, dt)
}
\arguments{
\item{m}{mu - extinction rate}

\item{l}{lambda - speciation rate}

\item{dt}{duration of the interval/period}
}
\description{
A function to calculate the probability of extinction for a paraclade.
}
