% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate2prob.R
\name{rate2prob}
\alias{rate2prob}
\title{A simple function to translate a rate into a probability.}
\usage{
rate2prob(x, dt)
}
\arguments{
\item{x}{rate}

\item{dt}{interval}
}
\value{
probability.
}
\description{
A simple function to translate a rate into a probability.
}
