% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remake_fosrec.R
\name{remake_fosrec}
\alias{remake_fosrec}
\title{Function to resample a fossil record from a simulated clade.}
\usage{
remake_fosrec(sim = sim, samp = function(t, n) {
    0.01
})
}
\arguments{
\item{sim}{sets the type parameterization of the rates of speciation and extinction. 'I' and 'II' specifies that the probabilities of extinction and seniority are calculated based on actual durations of intervals, whereas 'III' and 'IV' uses the differences between mid-points of intervals. 'I' and 'III' uses a basic implementation of probability of extinction as 1 - exp(-rate * dt). Models 'II' and 'IV' use a more detailed functional relationship between rate and probability, following Raup (1984):  (µ*((exp((lambda-µ)*dt))-1))/(lambda*((exp((lambda-µ)*dt)))-µ). Modeltype 'V' implements a temporally uninformed model, estimating the probabilities directly (or rather their logit).}

\item{samp}{function detailing sampling rate with arguments (t,n)}
}
\value{
The function returns an object of class CMR_model. This most important output is out$probfun which is a function for the posterior. This output can be fed directly into the sampler \link{MCMC_CMR}. It also returns \emph{Obs} and \emph{dts} as well as the full call (exlcluding default settings), \emph{call}
}
\description{
Function to resample a fossil record from a simulated clade.
}
