% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_time.R
\name{rescale_time}
\alias{rescale_time}
\title{Functino to rescale between time-scales.}
\usage{
rescale_time(scale1, scale2, x)
}
\arguments{
\item{scale1}{scale on which the times in x lie}

\item{scale2}{scale on which the times x should be changed into}

\item{x}{times of observations on scale1}
}
\value{
x_new - x on scale 2
}
\description{
This function takes the time-points x which belong to scale1 and rescales
them to scale 2. Scale1 and scale2 are two sets of ordered boundaries.
}
\details{
This function can be used to change a time on Gradstein 2004 to Gradstein 2012 for instance.
}
