% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampFosRec.R
\name{sampFosRec}
\alias{sampFosRec}
\title{Function to sample fossil from a lineage.}
\usage{
sampFosRec(orig, ext, samp)
}
\arguments{
\item{orig}{time of origination of the lineage}

\item{ext}{time of extinction of a lineage}

\item{samp}{how fossils are sampled. If given as a number it is interpreted as a proper rate (with units fossils / unit time). Can also be given as a function of time taking argument (t). Note that t is in absolute terms and not from time of origin of the lineage.}
}
\value{
an array of times at which the lineage is sampled
}
\description{
A simple function that yields a 'fossil record' of a lineage. Frequently used in order to simulate
fossil records for testing models, perform adequacy analyses or just for fun.
}
\examples{
Fos <- sampFosRec(orig = 2,ext = 12,samp = function(t){0.1+0.01*t})
Here a lineage originated at time 2 and went extinct at time 12 is sampled according to a linearly increasing samplin rate.
}
