% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doBML.R
\name{doBML}
\alias{doBML}
\title{Function to approximate Bayesian model probability using importance sampling.
The output logBML is approximating the Bayesian factor integral, and the support for one model over another, given the same data (i.e. the Bayes factor), can be approximated by exp(logBML[1] - logBML[2])}
\usage{
doBML(cmrfit, ndraws = 10000)
}
\arguments{
\item{cmrfit}{A fit of a model using Compadre. The function extracts the Bayesian probability density function and uses the last half of the chain.}

\item{ndraws}{number of samples to use to approximate the Bayesian model probability}
}
\description{
Function to approximate Bayesian model probability using importance sampling.
The output logBML is approximating the Bayesian factor integral, and the support for one model over another, given the same data (i.e. the Bayes factor), can be approximated by exp(logBML[1] - logBML[2])
}
