% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Foote_percap.R
\name{Foote_percap}
\alias{Foote_percap}
\title{Function to calculate Foote's rates of macroevolutionary change.}
\usage{
Foote_percap(Obs, dts = rep(1, dim(Obs)[2]))
}
\arguments{
\item{Obs}{matrix of observations (taxa by time)}

\item{dts}{vector of interval durations (defaults to 1)}
}
\value{
lambda,mu, bottomboundcros, topboundcros, rangethrough
}
\description{
Function to calculate Foote's rates of macroevolutionary change.
}
