% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_BayesCMR.R
\name{make_BayesCMR}
\alias{make_BayesCMR}
\title{Function to construct one clade CMR-model.}
\usage{
make_BayesCMR(Obs, dts = rep(1, dim(Obs)[2]), spec = ~1, ext = ~1,
  samp = ~1, data = NULL, pfix = 2, priorsNorm_Cov = c(0, 2),
  priorsNorm_Mus = rep(list(c(-4, 4)), 3), priorsUnif = c(-3, 3),
  replRE_1 = F)
}
\arguments{
\item{Obs}{a matrix with size \emph{number of taxa} by \emph{number of intervals (n)}. Each taxa has a row with 0's (unobserved) and 1's (observed) for each interval in the analysis. Oldest interval is first column.}

\item{dts}{a vector of interval durations. Defaults to a series of 1's if not supplied. Oldest interval is first entry.}

\item{data}{a data frame with putative drivers.}

\item{pfix}{a switch to select which approach is used to solve the identifiability problem in the model. If \emph{pfix = 1}, then the sampling rate in the first and the last intervals are assumed to be equal to the mean sampling rate for the whole period. If \emph{pfix = 2}, then the first two intervals, and the last two intervals have the same sampling rate. Defaults to pfix = 2.}

\item{priorsNorm_Cov}{sets the parameters for the normal prior for the impact of the drivers. Defaults to Norm(mu=0,sd=2).}

\item{priorsNorm_Mus}{sets the parameters for the normal prior for the (log) mean speciation, extinction and sampling rates. Given as a list of three entries with (mu,sd). Defaults to rep(list(c(-4,4)),3), i.e. all normal priors with mu=-4 and sd=3.}

\item{priorsUnif}{sets the parameters for the uniform prior for the log variances of the random efSpecTS. Same for all three variances, dunif(-3,3).}

\item{spec/ext/samp}{are formulas specifying the particular model. Use 'time' for temporally variable rates, 'div' for diversity dependence. Other drivers should have names as entries into \emph{data}}
}
\value{
The function returns an object of class CMR_model. This most important output is out$probfun which is a function for the posterior. This output can be fed directly into the sampler \link{MCMC_CMR}. It also returns \emph{Obs} and \emph{dts} as well as the full call (exlcluding default settings), \emph{call}
}
\description{
This function generates a structure containing the necessary functions for a CMR analysis of a fossil dataset.
}
\examples{
mod1 <- make.BayesCMR(Obs)
fit <- MCMC_CMR(mod1)
matplot(fit$Chain[,1:3],type="l")
This version implements the speciation and extinction rates using the length of the bins, not the
inter-bin midpoint difference.
}
