% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_BayesCMR_2clades.R
\name{make_BayesCMR_2clades}
\alias{make_BayesCMR_2clades}
\title{Function to construct two clade CMR-model.}
\usage{
make_BayesCMR_2clades(Obs1, Obs2, dts = rep(1, dim(Obs1)[2]), RE = c(T,
  T, T), DivDepSpecMatrix = matrix(rep(NA, 4), nrow = 2),
  DivDepExtMatrix = matrix(rep(NA, 4), nrow = 2), SpecTS_1 = NULL,
  ExtTS_1 = NULL, SmpTS_1 = NULL, SpecTS_2 = NULL, ExtTS_2 = NULL,
  SmpTS_2 = NULL, Driv_x_Div_Spec_1 = NULL, Driv_x_Div_Ext_1 = NULL,
  Driv_x_Div_Spec_2 = NULL, Driv_x_Div_Ext_2 = NULL, ...)
}
\arguments{
\item{Obs1}{observations clade 1: a matrix with size \emph{number of taxa} by \emph{number of intervals (n)}. Each taxa has a row with 0's (unobserved) and 1's (observed) for each interval in the analysis. Oldest interval is first column.}

\item{Obs2}{observations clade 1: a matrix with size \emph{number of taxa} by \emph{number of intervals (n)}. Each taxa has a row with 0's (unobserved) and 1's (observed) for each interval in the analysis. Oldest interval is first column.}

\item{dts}{a vector of interval durations. Defaults to series of 1's if not given.}

\item{DivDepSpecMatrix}{matrix of interactions between clades on speciation rates. Given as unique integers, starting at 1. For only diversity dependence of speciation rates within clades = matrix(c(1,NA,NA,2),nrow=2). For only between clade interactions = matrix(c(NA,1,2,NA),nrow=2). Similarly for DivDepExtMatrix, starting at the ma(DivDepSpecMatrix)+1}

\item{SpecTS_1/2}{drivers for speciation rates for clades 1/2}

\item{ExtTS_1/2}{drivers for extinction rates for clades 1/2}

\item{SmpTS_1/2}{drivers for sampling rates}

\item{Driv_x_Div_Spec_1/2}{whether or not speciation drivers interact with diversity in clade 1/2}

\item{Driv_x_Div_Spec_1/2}{whether or not extinction drivers interact with diversity in clade1/2. TRUE or FALSE of length equal to number of drivers.}
}
\value{
this function yields whatnow?
}
\description{
This function generates a structure containing the necessary functions for a
CMR analysis of two clades. In addition to drivers affecting both clades, their rates of
speciation and extinction can depend on the estimated diversity of both clades.
}
\details{
Both clades must be observed at each interval and these must be the same for the two clades.
}
\examples{
M1 <- make_BayesCMR_2clades(Obs1,Obs2,dts=rep(2,5))



}
