% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRates.R
\name{plotRates}
\alias{plotRates}
\title{Extract and plot macroevolutionary rates from a CMR model fit.}
\usage{
plotRates(
  cmrfit,
  max_ma = NULL,
  stages = NULL,
  logax = T,
  draws = 250,
  quantiles = c(0.025, 0.5, 0.975),
  plot = T,
  cladenames = NULL
)
}
\arguments{
\item{cmrfit}{A model fit, as output from \link{MCMC_CMR}}

\item{max_ma}{(optional) a maximum age for the oldest interval, if not given assumed to be sum of all interval durations in cmrfit}

\item{stages}{(optional) a data.frame of stages used for the analysis. Values max_ma,min_ma and colors are used to color the plot.}

\item{logax}{(T/F), use log y-axis for plotting}

\item{draws}{number of samples drawn from the MCMC chain (single number). Alternatively could be set of values indexing into the MCMC-chain.}

\item{plot}{(T/F) if plot figure, defaults to TRUE.}

\item{qauntiles}{gives the upper and lower quantiles for plotting bars around median, and the centre. Defaults to c(0.025,0.5,0.975).}
}
\value{
returns a list of three matrices with sampled rates; $SpecRates, $ExtRates and $SampRates.
}
\description{
Extract and plot macroevolutionary rates from a CMR model fit.
}
